Refactor VisualizationApp to use Visualization class
